stream/pyproject.toml
[project]
name = "stream"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "nadeem", email = "nadeemsangrasi903@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "openai-agents[litellm]>=0.0.16",
    "python-dotenv>=1.1.0",
]

[project.scripts]
stream = "stream:main"
agent = "stream.main:start"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
